+++ config/nvim/init.lua
@@ -1,5 +1,5 @@
-require("pkgs").setup()
-require("built_in_settings").setup()
-require("pkg_settings").setup()
-require("maps").setup()
-require("lsp").setup()
+require("aj.pkgs").setup()
+require("aj.built_in_settings").setup()
+require("aj.pkg_settings").setup()
+require("aj.maps").setup()
+require("aj.lsp").setup()
